Turned a size_t into an int, because it could be set to -1
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1532 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/ModelView/Dim3_Loader.cpp
@@ -234,7 +234,7 @@ bool LoadModel_Dim3(FileSpecifier& Spec, Model3D& Model, int WhichPass)
 		
 		// Add the bones, one by one;
 		// the bone stack's height is originally zero
-		size_t StackTop = -1;
+		int StackTop = -1;
 		for (size_t ib=0; ib<NumBones; ib++)
 		{		
 			// Scan down the bone stack to find a bone that's the parent of some unlisted bone;
